Make a sub folder for Feroese' original application, and rename my Monte Carloscript to "app.py"
