+++ app.py
@@ -1,379 +1,330 @@
-from scipy import interpolate
-import dash_table
-# import mydcc
-import dash_bootstrap_components as dbc
-from datetime import datetime as dt
+import numpy as np
+import pandas as pd
+
+pd.options.mode.chained_assignment = None  # default='warn'
+from scipy.integrate import odeint
+import matplotlib.pyplot as plt
+import plotly.express as px
 import dash
 import dash_core_components as dcc
 import dash_html_components as html
-import pandas as pd
-# import plotly.graph_objs as go
-import numpy as np
-pd.options.mode.chained_assignment = None  # default='warn'
-from scipy.integrate import odeint
 
 
 ############################################ the model ################################################
-
-
-def deriv(y, t, r0_y_interpolated, gamma, sigma, N, p_I_to_C, p_C_to_D, Beds):
+def deriv(y, t, N, p_immune, beta, gamma, delta, p_I_to_C, T_I_to_C, p_C_to_D, T_C_to_R, T_C_to_D):
     S, E, I, C, R, D = y
-
-    def betaa(t):
-        return I / (I + C) * (12 * p_I_to_C + 1/gamma * (1 - p_I_to_C)) + C / (I + C) * (
-            min(Beds(t), C) / (min(Beds(t), C) + max(0, C-Beds(t))) * (p_C_to_D * 7.5 + (1 - p_C_to_D) * 6.5) +
-            max(0, C-Beds(t)) / (min(Beds(t), C) + max(0, C-Beds(t))) * 1 * 1
-        )
-
-    def beta(t):
-        try:
-            return r0_y_interpolated[int(t)] / betaa(t) if not np.isnan(betaa(t)) else 0
-        except:
-            return r0_y_interpolated[-1] / betaa(t)
-            
-    dSdt = -beta(t) * I * S / N
-    dEdt = beta(t) * I * S / N - sigma * E
-    dIdt = sigma * E - 1/12.0 * p_I_to_C * I - gamma * (1 - p_I_to_C) * I
-    dCdt = 1/12.0 * p_I_to_C * I - 1/7.5 * p_C_to_D * \
-        min(Beds(t), C) - max(0, C-Beds(t)) - \
-        (1 - p_C_to_D) * 1/6.5 * min(Beds(t), C)
-    dRdt = gamma * (1 - p_I_to_C) * I + (1 - p_C_to_D) * \
-        1/6.5 * min(Beds(t), C)
-    dDdt = 1/7.5 * p_C_to_D * min(Beds(t), C) + max(0, C-Beds(t))
+    dSdt = -beta * I * S * (1 - p_immune) / N
+    dEdt = beta * I * S * (1 - p_immune) / N - delta * E
+    dIdt = delta * E - 1 / T_I_to_C * p_I_to_C * I - gamma * (1 - p_I_to_C) * I
+    dCdt = 1 / T_I_to_C * p_I_to_C * I - 1 / T_C_to_D * p_C_to_D * C - (1 - p_C_to_D) * 1 / T_C_to_R * C
+    dRdt = gamma * (1 - p_I_to_C) * I + (1 - p_C_to_D) * 1 / T_C_to_R * C
+    dDdt = 1 / T_C_to_D * p_C_to_D * C
     return dSdt, dEdt, dIdt, dCdt, dRdt, dDdt
 
 
-
-gamma = 1.0/9.0
-sigma = 1.0/3.0
-
-def logistic_R_0(t, R_0_start, k, x0, R_0_end):
-    return (R_0_start-R_0_end) / (1 + np.exp(-k*(-t+x0))) + R_0_end
-
-def Model(initial_cases, initial_date, N, beds_per_100k, R_0_start, k, x0, R_0_end, p_I_to_C, p_C_to_D, s, r0_y_interpolated=None):
-    days = 360
-    def beta(t):
-        return logistic_R_0(t, R_0_start, k, x0, R_0_end) * gamma
-    
-    def Beds(t):
-        beds_0 = beds_per_100k / 100_000 * N
-        return beds_0 + s*beds_0*t  # 0.003
-
-
-    diff = int((np.datetime64("2020-01-01") - np.datetime64(initial_date)) / np.timedelta64(1, "D"))
-    if diff > 0:
-        r0_y_interpolated = [r0_y_interpolated[0] for _ in range(diff-1)] + r0_y_interpolated
-    elif diff < 0:
-        r0_y_interpolated = r0_y_interpolated[(-diff):]
-
-    last_date = np.datetime64(initial_date) + np.timedelta64(days-1, "D")
-    missing_days_r0 = int((last_date - np.datetime64("2020-09-01")) / np.timedelta64(1, "D"))
-    r0_y_interpolated += [r0_y_interpolated[-1] for _ in range(missing_days_r0+1)]
-
-    y0 = N-initial_cases, initial_cases, 0.0, 0.0, 0.0, 0.0
-    t = np.linspace(0, days, days)
-    print(t)
-    ret = odeint(deriv, y0, t, args=(r0_y_interpolated,
-                                        gamma, sigma, N, p_I_to_C, p_C_to_D, Beds))
+def Model(initial_cases, initial_date, N, p_immune, beta, gamma, delta, p_I_to_C, T_I_to_C, p_C_to_D, T_C_to_R,
+          T_C_to_D):
+    days = 365
+    y0 = N - initial_cases, initial_cases, 0.0, 0.0, 0.0, 0.0
+    t = np.linspace(1, days, days)  # ORN changed start = 0 to start = 1 for integer values
+    #    print(t)
+    ret = odeint(deriv, y0, t, args=(N, p_immune, beta, gamma, delta, p_I_to_C, T_I_to_C, p_C_to_D, T_C_to_R, T_C_to_D))
+    # todo: make sure to find a way to implement the critical community size truncating - Actually, Steve said no need
     S, E, I, C, R, D = ret.T
-    R_0_over_time = r0_y_interpolated
-    total_CFR = [0] + [100 * D[i] / sum(sigma*E[:i]) if sum(
-        sigma*E[:i]) > 0 else 0 for i in range(1, len(t))]
-    daily_CFR = [0] + [100 * ((D[i]-D[i-1]) / ((R[i]-R[i-1]) + (D[i]-D[i-1]))) if max(
-        (R[i]-R[i-1]), (D[i]-D[i-1])) > 10 else 0 for i in range(1, len(t))]
+    # note this is actually IFR: Dead / total_infected
+    total_CFR = [0] + [100 * D[i] / sum(delta * E[:i]) if sum(
+        delta * E[:i]) > 0 else 0 for i in range(1, len(t))]
+    daily_CFR = [0] + [100 * ((D[i] - D[i - 1]) / ((R[i] - R[i - 1]) + (D[i] - D[i - 1]))) if max(
+        (R[i] - R[i - 1]), (D[i] - D[i - 1])) > 10 else 0 for i in range(1, len(t))]
+    return t, S, E, I, C, R, D, total_CFR, daily_CFR
+
+
+############### initialization ###############
+
+
+N = 100000  # size of initial population
+initial_cases = 10  # initial number of Exposed. I chose 10 to stay away from the critical community size at the onset of the epidemic.
+
+num_reps = 50  # for Monte Carlo
+
+disease = 'Covid_19'  # supported options: 'Covid_19', 'Influenza_seasonal', or 'deterministic_test'
+print(disease)
+# times below are assumed to be common for Covid-19 and influenza seasonal
+# todo: if supporting more disease, and they don't share the same timing, move those into the "if disease " above.
+T_I_to_C = np.random.uniform(4.5, 12.5, num_reps).round(2)  # time to transition from Infected to Critical [days]
+T_C_to_D = np.random.uniform(5, 9, num_reps).round(2)  # time to transition from Critical to Dead [days]
+T_C_to_R = np.random.uniform(12, 16, num_reps).round(2)  # time to transition from Critical to Recovered [days]
+
+# The section below is intended to be modular, so we can easily add other disease model.
+if disease == 'Covid_19':
+    R_0 = np.random.uniform(1.94, 5.7, num_reps).round(2)  # Initial R0
+    p_immune = np.random.uniform(5 / 100, 7 / 100, num_reps).round(3)  # portion of immuned population
+    T_infectious = np.random.uniform(8, 12, num_reps).round(2)  # infectious period [days]
+    T_incubation = np.random.uniform(3, 8, num_reps).round(
+        2)  # incubation period [days] - not contagious during this time
+    gamma = 1.0 / T_infectious
+    delta = 1.0 / T_incubation
+    beta = R_0 / T_infectious
+    p_I_to_C = np.random.uniform(3 / 100, 8 / 100, num_reps).round(
+        3)  # overall probability of transition from Infected to Critical
+    p_C_to_D = np.random.uniform(25 / 100, 40 / 100, num_reps).round(
+        3)  # overall probability of transition from Infected to Critical
+elif disease == 'Influenza_seasonal':
+    R_0 = np.random.uniform(0.9, 2.1, num_reps).round(2)  # Initial R0
+    p_immune = np.random.uniform(15 / 100, 25 / 100, num_reps).round(3)  # portion of immuned population
+    T_infectious = np.random.uniform(3, 10, num_reps).round(2)  # infectious period [days]
+    T_incubation = np.random.uniform(1, 4, num_reps).round(
+        2)  # incubation period [days] - not contagious during this time
+    gamma = 1.0 / T_infectious
+    delta = 1.0 / T_incubation
+    beta = R_0 / T_infectious
+    p_I_to_C = np.random.uniform(0.25 / 100, 0.75 / 100, num_reps).round(
+        3)  # overall probability of transition from Infected to Critical
+    p_C_to_D = np.random.uniform(33 / 100, 48 / 100, num_reps).round(
+        3)  # overall probability of transition from Infected to Critical
+elif disease == 'deterministic_test':
+    R_0 = np.random.uniform(3, 3, num_reps).round(2)  # Initial R0
+    p_immune = np.random.uniform(0 / 100, 0 / 100, num_reps).round(3)  # portion of immuned population
+    T_infectious = np.random.uniform(9, 9, num_reps).round(2)  # infectious period [days]
+    T_incubation = np.random.uniform(3, 3, num_reps).round(
+        2)  # incubation period [days] - not contagious during this time
+    gamma = 1.0 / T_infectious
+    delta = 1.0 / T_incubation
+    beta = R_0 / T_infectious
+    p_I_to_C = np.random.uniform(5 / 100, 5 / 100, num_reps).round(
+        3)  # overall probability of transition from Infected to Critical
+    p_C_to_D = np.random.uniform(30 / 100, 30 / 100, num_reps).round(
+        3)  # overall probability of transition from Infected to Critical
+    T_I_to_C = np.random.uniform(12, 12, num_reps).round(2)  # time to transition from Infected to Critical [days]
+    T_C_to_D = np.random.uniform(7.5, 7.5, num_reps).round(2)  # time to transition from Critical to Dead [days]
+    T_C_to_R = np.random.uniform(6.5, 6.5, num_reps).round(2)  # time to transition from Critical to Recovered [days]
+else:
+    print('undefined disease!')
+
+initial_date = 0
+
+############### Monte Carlo ###############
+print('start Monte-Carlo')
+df_results = pd.DataFrame({})  # initiate as an empty DataFrame so I can append to it in the Monte Carlo loop
+df_I_vs_t = pd.DataFrame([])  # initiate as an empty DataFrame so I can append to it in the Monte Carlo loop
+I_vs_t = []
+df_C_vs_t = pd.DataFrame([])  # initiate as an empty DataFrame so I can append to it in the Monte Carlo loop
+C_vs_t = []
+df_D_vs_t = pd.DataFrame([])  # initiate as an empty DataFrame so I can append to it in the Monte Carlo loop
+D_vs_t = []
+df_IFR_vs_t = pd.DataFrame([])  # initiate as an empty DataFrame so I can append to it in the Monte Carlo loop
+IFR_vs_t = []
+for i in range(num_reps):
+    time, S, E, I, C, R, D, total_CFR, daily_CFR = Model(initial_cases, initial_date, N, p_immune[i], beta[i], gamma[i],
+                                                         delta[i], p_I_to_C[i], T_I_to_C[i], p_C_to_D[i], T_C_to_R[i],
+                                                         T_C_to_D[i])
+    df_temp = pd.DataFrame(index=[i], data={'R_0': R_0[i], 'immune %': p_immune[i] * 100,
+                                            'T_infectious': T_infectious[i], 'T_incubation': T_incubation[i],
+                                            'p_I_to_C%': p_I_to_C[i] * 100, 'p_C_to_D%': p_C_to_D[i] * 100,
+                                            'T_I_to_C': T_I_to_C[i], 'T_C_to_R': T_C_to_R[i], 'T_C_to_D': T_C_to_D[i],
+                                            'peak E': max(E), 'peak I': max(I), 'peak R': max(R),
+                                            'peak C': max(C), 'max D': max(D), 'total IFR': max(total_CFR)})
+    df_results = df_results.append(df_temp)
+    # plot progress to terminal, to avoid the feeling it got stuck...
+    if round(i/25, 0) == (i/25):
+        print('in MC round ', i, 'out of ', num_reps)
+
+    # thanks to
+    # https://stackoverflow.com/questions/31674557/how-to-append-rows-in-a-pandas-dataframe-in-a-for-loop for this method
+    IFR_vs_t.append(total_CFR)
+    I_vs_t.append(I)
+    D_vs_t.append(D)
+    C_vs_t.append(C)
+    df_IFR_vs_t = pd.DataFrame(IFR_vs_t)  # compartment value for each day is appended as new row
+    df_I_vs_t = pd.DataFrame(I_vs_t)  # compartment value for each day is appended as new row
+    df_C_vs_t = pd.DataFrame(C_vs_t)  # compartment value for each day is appended as new row
+    df_D_vs_t = pd.DataFrame(D_vs_t)  # compartment value for each day is appended as new row
+
+########## print some results to the Terminal ###################
+# Set it to None to display all columns in the dataframe,
+# but actually print to terminal only if <30 rows (30 being arbitrary, used for testing)
+if num_reps < 30:
+    pd.set_option('display.max_columns', None)
+    print(df_results)
+else:
+    print('NOT printing df_results to Terminal, as it is deemed too long')
+print('min / max I')
+print(min(df_results['peak I']), max(df_results['peak I']))
+print('min / max C')
+print(min(df_results['peak C']), max(df_results['peak C']))
+print('min / max D')
+print(min(df_results['max D']), max(df_results['max D']))
+print('min(IFR), max(IFR)')
+print('min / max R')
+print(min(df_results['peak R']), max(df_results['peak R']))
+print(min(df_results['total IFR']), max(df_results['total IFR']))
+
+#############  try using dash  ###############
+
+
+app = dash.Dash()
+#  plot using ploty express, with the intention to transfer it to DCC as explained in https://dash.plotly.com/dash-core-components/graph
+#  could not make it run with the regular dcc.Graph( ...  id='example-graph', ... figure={ ...  'data': [
+#  in that case it worked for a single line but I couldn't figure out how to pull all lines from the DataFrame. It kept failing to load the web page.
+# subplots: by https://panel.holoviz.org/reference/panes/Plotly.html
+fig1 = px.line(df_I_vs_t[1:20].T,
+               title="Number of Infected vs. time, per Monte Carlo run (1st 20)",
+               width=600, height=400,
+               labels={  # replaces default labels by column name
+                   # ... thanks to https://plotly.com/python/styling-plotly-express/
+                   "index": "time [days]", "value": "Number of Infected"
+               },
+               )
+fig1.layout.update(showlegend=False)  # eliminate legends as they can be very long (one for each Monte Carlo run)
+
+fig2 = px.line(df_C_vs_t[1:20].T,
+               title="Number of Critical vs. time, per Monte Carlo run (1st 20)",
+               width=600, height=400,
+               labels={  # replaces default labels by column name
+                   # ... thanks to https://plotly.com/python/styling-plotly-express/
+                   "index": "time [days]", "value": "Number of Critical"
+               },
+               )
+fig2.layout.update(showlegend=False)  # eliminate legends as they can be very long (one for each Monte Carlo run)
+
+fig3 = px.line(df_D_vs_t[1:20].T,
+               title="Number of Dead vs. time, per Monte Carlo run (1st 20)",
+               width=600, height=400,
+               labels={  # replaces default labels by column name
+                   # ... thanks to https://plotly.com/python/styling-plotly-express/
+                   "index": "time [days]", "value": "Number of Dead"
+               },
+               )
+fig3.layout.update(showlegend=False)  # eliminate legends as they can be very long (one for each Monte Carlo run)
+
+fig4 = px.histogram(df_results, x="total IFR",
+                    title="Histogram of IFR (Infected / Dead) over all Monte Carlo runs",
+                    width=600, height=400,
+                    )
+fig5 = px.histogram(df_results, x="peak C",
+                    title="Histogram of maximum number of Critical over all Monte Carlo runs",
+                    width=600, height=400,
+                    )
+fig6 = px.histogram(df_results, x="max D",
+                    title="Histogram of final Dead number over all Monte Carlo runs",
+                    width=600, height=400,
+                    )
+
+fig7 = px.line(df_I_vs_t.T.apply(np.median, axis=1),
+               title="Median of Infected vs. time, over all Monte Carlo run",
+               width=600, height=400,
+               labels={  # replaces default labels by column name
+                   # ... thanks to https://plotly.com/python/styling-plotly-express/
+                   "index": "time [days]", "value": "Median of Infected"
+               },
+               )
+fig7.layout.update(showlegend=False)  # eliminate legends
+fig8 = px.line(df_C_vs_t.T.apply(np.median, axis=1),
+               title="Median of Critical vs. time, over all Monte Carlo run",
+               width=600, height=400,
+               labels={  # replaces default labels by column name
+                   # ... thanks to https://plotly.com/python/styling-plotly-express/
+                   "index": "time [days]", "value": "Median of Critical"
+               },
+               )
+fig8.layout.update(showlegend=False)  # eliminate legends
+
+fig9 = px.line(df_D_vs_t.T.apply(np.median, axis=1),
+               title="Median of Dead vs. time, over all Monte Carlo run",
+               width=600, height=400,
+               labels={  # replaces default labels by column name
+                   # ... thanks to https://plotly.com/python/styling-plotly-express/
+                   "index": "time [days]", "value": "Median of Dead"
+               },
+               )
+fig9.layout.update(showlegend=False)  # eliminate legends
+
+fig10 = px.line(df_IFR_vs_t.T.apply(np.median, axis=1),
+               title="Median of IFR vs. time, over all Monte Carlo run",
+               width=600, height=400,
+               labels={  # replaces default labels by column name
+                   # ... thanks to https://plotly.com/python/styling-plotly-express/
+                   "index": "time [days]", "value": "Median of IFR"
+               },
+               )
+fig10.layout.update(showlegend=False)  # eliminate legends
 
 
 
-    dates = pd.date_range(start=np.datetime64(initial_date), periods=days, freq="D")
+app.layout = html.Div(children=[
+    html.H1(children='SEIR Model, Monte Carlo results'),
+    html.H2(children=disease),
+    html.Div(children=['Population size =', N]),
+    html.Div(children=['initial number of Infected =', initial_cases]),
+    html.Div(children=['number of Monte Carlo Runs =', num_reps]),
+    html.Div(children='other important parameters, all with uniformly-distributed ranges set in the code:'
+                      ' % of immuned population, R0, T_incubation, T_infected, '
+                      'P(I2C), P(C2D), PT(I2C), T(C2R), T(C2D)'),
+    html.Div(children='''
+        I wish I could add an interactive section to present and adjust all parameters, 
+        not only initial population... 
+        and to find a way to arrange the figure side by side rather than one after the other...  
+    '''),
+    dcc.Graph(
+        id='fig1',
+        figure=fig1,
+    ),
+    dcc.Graph(
+        id='fig2',
+        figure=fig2,
+    ),
+    dcc.Graph(
+        id='fig3',
+        figure=fig3,
+    ),
 
-    return dates, S, E, I, C, R, D, R_0_over_time, total_CFR, daily_CFR, [Beds(i) for i in range(len(t))]
+    dcc.Graph(
+        id='fig4',
+        figure=fig4,
+    ),
+    dcc.Graph(
+        id='fig5',
+        figure=fig5,
+    ),
+    dcc.Graph(
+        id='fig6',
+        figure=fig6,
+     ),
+    html.Div(children='Please take the median graphs below with a grain of salt,'
+                      ' as the median will reflect the mid-range of the parameters distribution, '
+                      'and to an extent will negate the uniform distribution we chose for the analysis'),
+    dcc.Graph(
+        id='fig10',
+        figure=fig10,
+    ),
+    dcc.Graph(
+        id='fig7',
+        figure=fig7,
+    ),
+    dcc.Graph(
+        id='fig8',
+        figure=fig8,
+    ),
+    dcc.Graph(
+        id='fig9',
+        figure=fig9,
+    )
 
-
-############################################ t
-# he dash app layout ################################################
-external_stylesheets = [dbc.themes.BOOTSTRAP]
-
-app = dash.Dash(__name__, external_stylesheets=external_stylesheets)
-app.title = "CoSim"
-
-
-# these are the controls where the parameters can be tuned.
-# They are not placed on the screen here, we just define them.
-# Each separate input (e.g. a slider for the fatality rate) is placed
-# in its own "dbc.FormGroup" and gets a "dbc.Label" where we put its name.
-# The sliders use the predefined "dcc.Slider"-class, the numeric inputs
-# use "dbc.Input", etc., so we don't have to tweak anything ourselves.
-# The controls are wrappen in a "dbc.Card" so they look nice.
-controls = dbc.Card(
-    [
-        dbc.FormGroup(
-            [
-                dbc.Label('Date of first infection'),
-                html.Br(),
-                dcc.DatePickerSingle(
-                    day_size=39,  # how big the date picker appears
-                    display_format="DD.MM.YYYY",
-                    date='2020-01-01',
-                    id='initial_date',
-                    min_date_allowed=dt(2020, 12, 1),
-                    max_date_allowed=dt(2020, 5, 31),
-                    initial_visible_month=dt(2020, 1, 15),
-                    placeholder="test"
-                ),
-            ]
-        ),
-        dbc.FormGroup(
-            [
-                dbc.Label("Initial Cases"),
-                dbc.Input(
-                    id="initial_cases", type="number", placeholder="initial_cases",
-                    min=1, max=1_000_000, step=1, value=10,
-                )
-            ]
-        ),
-
-        dbc.FormGroup(
-            [
-                dbc.Label("Population"),
-                dbc.Input(
-                    id="population", type="number", placeholder="population",
-                    min=10_000, max=1_000_000_000, step=10_000, value=80_000_000,
-                )
-            ]
-        ),
-        dbc.FormGroup(
-            [
-                dbc.Label('ICU beds per 100k people'),
-                dbc.Input(
-                    id="icu_beds", type="number", placeholder="ICU Beds per 100k",
-                    min=0.0, max=100.0, step=0.1, value=34.0,
-                ),
-            ]
-        ),
-        dbc.FormGroup(
-            [
-                dbc.Label('Probability of going to ICU when infected (%)'),
-                html.Br(),
-                dcc.Slider(
-                    id='p_I_to_C',
-                    min=0.01,
-                    max=100.0,
-                    step=0.01,
-                    value=5.0,
-                    tooltip={'always_visible': True, "placement": "bottom"}
-                ),
-            ]
-        ),
-        dbc.FormGroup(
-            [
-                dbc.Label('Probability of dying in ICU (%)'),
-                dcc.Slider(
-                    id='p_C_to_D',
-                    min=0.01,
-                    max=100.0,
-                    step=0.01,
-                    value=50.0,
-                    tooltip={'always_visible': True, "placement": "bottom"}
-                ),
-            ]
-        ),
-        # this is the input where the R value can be changed over time.
-        # It is implemented as a table where the date is in the first column,
-        # and users can change the R value on that date in the second column.
-        dbc.FormGroup(
-            [
-                dbc.Label('Reproduction rate R over time'),
-                dash_table.DataTable(
-                    id='r0_table',
-                    columns=[
-                        {"name": "Date", "id": "Date"},
-                        {"name": "R value", "id": "R value",
-                         "editable": True, "type": "numeric"},
-                    ],
-                    data=[
-                        {
-                            "Date": i[0],
-                            "R value": i[1],
-                        }
-                        for i in [("2020-01-01", 3.2), ("2020-02-01", 2.9), ("2020-03-01", 2.5), ("2020-04-01", 0.8), ("2020-05-01", 1.1), ("2020-06-01", 2.0), ("2020-07-01", 2.1), ("2020-08-01", 2.2), ("2020-09-01", 2.3)]
-                    ],
-                    style_cell_conditional=[
-                        {'if': {'column_id': 'Date'},
-                         'width': '5px'},
-                        {'if': {'column_id': 'R value'},
-                         'width': '10px'},
-                    ],
-                    style_cell={'textAlign': 'left',
-                                'fontSize': 16, 'font-family': 'Helvetica'},
-                    style_header={
-                        'backgroundColor': 'white',
-                        'fontWeight': 'bold'
-                    },
-
-                ),
-            ]
-        ),
-        dbc.Button("Apply", id="submit-button-state",
-                   color="primary", block=True)
-    ],
-    body=True,
-)
-
-# layout for the whole page
-app.layout = dbc.Container(
-    [
-        # first, a jumbotron for the description and title
-        dbc.Jumbotron(
-            [
-                dbc.Container(
-                    [
-                        html.H1("CoSim", className="display-3"),
-                        html.P(
-                            "Interactively simulate different Coronavirus scenarios. ",
-                            className="lead",
-                        ),
-                        html.Hr(className="my-2"),
-                        dcc.Markdown('''
-
-                            You can freely tune the date the first infection occured, the total population, the number of ICU
-                            beds per 100k people (you can look the numbers up [here](https://en.wikipedia.org/wiki/List_of_countries_by_hospital_beds#Numbers)),
-                            the probability of needing intensive care and the probability of dying under intensive care.
-                            You can also change the reproduction rate R over time to simulate lockdowns, social distancing, a potential second wave, etc.
-
-                            Read [this article](https://towardsdatascience.com/infectious-disease-modelling-part-i-understanding-sir-28d60e29fdfc) 
-                            if you want to know more about the various parameters, 
-                            and [this one](https://towardsdatascience.com/infectious-disease-modelling-beyond-the-basic-sir-model-216369c584c4) 
-                            if you want to learn about the exact model used here.
-                            '''
-                                     )
-                    ],
-                    fluid=True,
-                )
-            ],
-            fluid=True,
-            className="jumbotron bg-white text-dark"
-        ),
-        # now onto the main page, i.e. the controls on the left
-        # and the graphs on the right.
-        dbc.Row(
-            [
-                # here we place the controls we just defined,
-                # and tell them to use up the left 3/12ths of the page.
-                dbc.Col(controls, md=3),
-                # now we place the graphs on the page, taking up
-                # the right 9/12ths.
-                dbc.Col(
-                    [
-                        # the main graph that displays coronavirus over time.
-                        dcc.Graph(id='main_graph'),
-                        # the graph displaying the R values the user inputs over time.
-                        dcc.Graph(id='r0_graph'),
-                        # the next two graphs don't need as much space, so we
-                        # put them next to each other in one row.
-                        dbc.Row(
-                            [
-                                # the graph for the fatality rate over time.
-                                dbc.Col(dcc.Graph(id='cfr_graph'), md=6),
-                                # the graph for the daily deaths over time.
-                                dbc.Col(dcc.Graph(id="deaths_graph"), md=6)
-
-                            ]
-                        ),
-                    ],
-                    md=9
-                ),
-            ],
-            align="top",
-        ),
-    ],
-    # fluid is set to true so that the page reacts nicely to different sizes etc.
-    fluid=True,
-)
-
-
-
-############################################ the dash app callbacks ################################################
-
-
-@app.callback(
-    [dash.dependencies.Output('main_graph', 'figure'),
-     dash.dependencies.Output('cfr_graph', 'figure'),
-     dash.dependencies.Output('r0_graph', 'figure'),
-     dash.dependencies.Output('deaths_graph', 'figure'),
-     ],
-     
-    [dash.dependencies.Input('submit-button-state', 'n_clicks')],
-
-    [dash.dependencies.State('initial_cases', 'value'),
-     dash.dependencies.State('initial_date', 'date'),
-     dash.dependencies.State('population', 'value'),
-     dash.dependencies.State('icu_beds', 'value'),
-     dash.dependencies.State('p_I_to_C', 'value'),
-     dash.dependencies.State('p_C_to_D', 'value'),
-     dash.dependencies.State('r0_table', 'data'),
-     dash.dependencies.State('r0_table', 'columns')
-     ]
-)
-
-
-def update_graph(_, initial_cases, initial_date, population, icu_beds, p_I_to_C, p_C_to_D, r0_data, r0_columns):
-    
-    last_initial_date, last_population, last_icu_beds, last_p_I_to_C, last_p_C_to_D = "2020-01-15", 1_000_000, 5.0, 5.0, 50.0
-    if not (initial_date and population and icu_beds and p_I_to_C and p_C_to_D):
-        initial_date, population, icu_beds, p_I_to_C, p_C_to_D = last_initial_date, last_population, last_icu_beds, last_p_I_to_C, last_p_C_to_D
-
-
-    r0_data_x = [datapoint["Date"] for datapoint in r0_data]
-    r0_data_y = [datapoint["R value"] if ((not np.isnan(datapoint["R value"])) and (datapoint["R value"] >= 0))  else 0 for datapoint in r0_data]
-    f = interpolate.interp1d([0, 1, 2, 3, 4, 5, 6, 7, 8], r0_data_y, kind='linear')
-    r0_x_dates = pd.date_range(start=np.datetime64("2020-01-01"), end=np.datetime64("2020-09-01"), freq="D")
-    r0_y_interpolated = f(np.linspace(0, 8, num=len(r0_x_dates))).tolist()
-
-    dates, S, E, I, C, R, D, R_0_over_time, total_CFR, daily_CFR, B = Model(initial_cases, initial_date, population, icu_beds, 3.0, 0.01, 50, 2.3, float(p_I_to_C)/100, float(p_C_to_D)/100, 0.001, r0_y_interpolated)
-
-    return {  # return graph for compartments, graph for fatality rates, graph for reproduction rate, and graph for deaths over time
-        'data': [
-            {'x': dates, 'y': Sdcc.Graph, 'type': 'line', 'name': 'susceptible'},
-            {'x': dates, 'y': E.astype(int), 'type': 'line', 'name': 'exposed'},
-            {'x': dates, 'y': I.astype(int), 'type': 'line', 'name': 'infected'},
-            {'x': dates, 'y': C.astype(int), 'type': 'line', 'name': 'critical'},
-            {'x': dates, 'y': R.astype(int), 'type': 'line', 'name': 'recovered'},
-            {'x': dates, 'y': D.astype(int), 'type': 'line', 'name': 'dead'},
-        ],
-        'layout': {
-            'title': 'Compartments over time'
-        }
-        }, {
-        'data': [
-            {'x': dates, 'y': daily_CFR, 'type': 'line',
-                'name': 'daily'},
-            {'x': dates, 'y': total_CFR, 'type': 'line',
-                'name': 'total'}
-        ],
-        'layout': {
-            'title': 'Fatality rate over time (%)',
-            }
-        }, {
-        'data': [
-            {'x': dates, 'y': R_0_over_time, 'type': 'line', 'name': 'susceptible'}
-        ],
-        'layout': {
-            'title': 'Reproduction Rate R over time',
-            }
-        }, {
-        'data': [
-            {'x': dates, 'y': [0] + [D[i]-D[i-1] for i in range(1, len(dates))], 'type': 'line', 'name': 'total'},
-            {'x': dates, 'y': [0] + [max(0, C[i-1]-B[i-1]) for i in range(1, len(dates))], 'type': 'line', 'name': 'due to overcapacity'}
-        ],
-        'layout': {
-            'title': 'Deaths per day'
-        }
-        }
-
-
-
-
-
+    ###   keep the section below commented out to see what did work and what had not (which id doubly commented out)
+    # dcc.Graph(
+    #     id='example-graph',
+    #     figure={
+    #         'data': [
+    #             # {'y': df_C_vs_t.T, 'type': 'line'}
+    #             {'y': df_C_vs_t.T[2], 'type': 'line'},
+    #             {'y': df_C_vs_t.T[3], 'type': 'line'}
+    #         ],
+    #         'layout': {
+    #             'title': disease
+    #         }
+    #     }
+    # )
+])
 
 if __name__ == '__main__':
     app.run_server(debug=True)
